backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pinecone-client==2.2.4
langchain==0.0.335
openai==1.3.5
pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.4.2
anthropic==0.15.0
supabase==1.0.3
